+++ rope.h
@@ -1,12 +1,15 @@
+#ifndef ROPE_H
+# define ROPE_H
+
 typedef struct s_splay_tree	t_splay_tree;
 struct	s_splay_tree
 {
-	void	*value;
+	void			*value;
 	t_splay_tree	*left;
 	t_splay_tree	*right;
 };
 
-typedef enum	e_splay_direction
+typedef enum e_splay_direction
 {
 	SPLAY_ROOT = 0x500,
 	SPLAY_LEFT,
@@ -47,16 +50,20 @@ t_splay_tree	*splay(t_splay_path *path);
 t_splay_path	*splay_path_left(t_splay_path *path);
 t_splay_path	*splay_path_right(t_splay_path *path);
 
-#define ROPE_NOWEIGHT ((void*)-1L)
+# define ROPE_NOWEIGHT ((void*)-1L)
 
-typedef t_splay_tree	t_rope;
+typedef struct s_splay_tree	t_rope;
 
-t_rope	*rope_create(char *left, char *right);
-t_rope	*rope_concat(t_rope *left, t_rope *right);
-int		rope_weight(t_rope *rope);
-char	rope_index(t_rope *rope, int index);
-char	rope_index_with_path(t_rope *rope, int index, t_splay_path **path);
-int		rope_length(t_rope *rope);
-void	rope_split(t_rope *rope, int index, t_rope **left, t_rope **right);
-t_rope	*rope_insert(t_rope *rope, int index, t_rope *right);
-t_rope	*rope_delete(t_rope *rope, int begin, int end);
+t_rope			*rope_create(char *left, char *right);
+t_rope			*rope_concat(t_rope *left, t_rope *right);
+int				rope_weight(t_rope *rope);
+char			rope_index(t_rope *rope, int index);
+char			rope_index_with_path(
+					t_rope *rope, int index, t_splay_path **path);
+int				rope_length(t_rope *rope);
+void			rope_split(
+					t_rope *rope, int index, t_rope **left, t_rope **right);
+t_rope			*rope_insert(t_rope *rope, int index, t_rope *right);
+t_rope			*rope_delete(t_rope *rope, int begin, int end);
+
+#endif
